+++ src/main/java/net/crypticverse/betterbiomes/util/BetterBiomesTags.java
@@ -1,10 +1,13 @@
 package net.crypticverse.betterbiomes.util;
 
+import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.BiomeTags;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.block.Block;
 
 import net.crypticverse.betterbiomes.BetterBiomes;
@@ -34,5 +37,14 @@ public class BetterBiomesTags {
             return BlockTags.create(new ResourceLocation("forge", name));
         }
 
+    }
+
+    public static class Biomes {
+        public static final TagKey<Biome> IS_THIN = tag("is_thin");
+
+
+        private static TagKey<Biome> tag(String name) {
+            return TagKey.create(Registries.BIOME, new ResourceLocation(BetterBiomes.MOD_ID, name));
+        }
     }
 }