+++ src/main/java/net/crypticverse/betterbiomes/world/BetterBiomeModifiers.java
@@ -1,10 +1,17 @@
 package net.crypticverse.betterbiomes.world;
 
+import com.bookkeepersmc.notebook.common.world.BiomeRegistry;
+import com.bookkeepersmc.notebook.common.world.SelectBiomes;
+import dev.architectury.registry.level.biome.BiomeModifications;
+import net.crypticverse.betterbiomes.util.BetterBiomesTags;
+import net.crypticverse.betterbiomes.world.biome.BetterBiomesBiomes;
 import net.minecraft.core.HolderSet;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BootstapContext;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.BiomeTags;
+import net.minecraft.world.level.biome.Biomes;
 import net.minecraft.world.level.levelgen.GenerationStep;
 
 import net.minecraftforge.common.Tags;
@@ -16,6 +23,7 @@ import net.crypticverse.betterbiomes.BetterBiomes;
 
 public class BetterBiomeModifiers {
     public static final ResourceKey<BiomeModifier> ADD_TREE_MAPLE = registerKey("add_tree_maple");
+    public static final ResourceKey<BiomeModifier> ADD_TREE_THIN = registerKey("add_tree_thin");
 
     public static void bootstrap(BootstapContext<BiomeModifier> context) {
         var placedFeatures = context.lookup(Registries.PLACED_FEATURE);
@@ -25,6 +33,7 @@ public class BetterBiomeModifiers {
                 biomes.getOrThrow(Tags.Biomes.IS_DENSE_OVERWORLD),
                 HolderSet.direct(placedFeatures.getOrThrow(BetterBiomePlacedFeatures.MAPLE_PLACED_KEY)),
                 GenerationStep.Decoration.VEGETAL_DECORATION));
+
     }