+++ src/main/java/net/crypticverse/betterbiomes/world/biome/BetterBiomesBiomes.java
@@ -3,6 +3,7 @@ package net.crypticverse.betterbiomes.world.biome;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.worldgen.BiomeDefaultFeatures;
 import net.minecraft.data.worldgen.BootstapContext;
+import net.minecraft.data.worldgen.placement.VegetationPlacements;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.EntityType;
@@ -15,9 +16,11 @@ import net.crypticverse.betterbiomes.world.BetterBiomePlacedFeatures;
 
 public class BetterBiomesBiomes {
     public static final ResourceKey<Biome> MAPLE_FOREST = register("maple_forest");
+    public static final ResourceKey<Biome> THIN_BIRCH_FOREST = register("thin_birch_forest");
 
     public static void bootstrap(BootstapContext<Biome> context) {
         context.register(MAPLE_FOREST, mapleForest(context));
+        context.register(THIN_BIRCH_FOREST, thinForest(context));
     }
     public static void globalOverworldGeneration(BiomeGenerationSettings.Builder builder) {
         BiomeDefaultFeatures.addDefaultCarversAndLakes(builder);
@@ -47,7 +50,7 @@ public class BetterBiomesBiomes {
         BiomeDefaultFeatures.addDefaultOres(biomeBuilder);
         BiomeDefaultFeatures.addPlainVegetation(biomeBuilder);
 
-        biomeBuilder.addFeature(GenerationStep.Decoration.VEGETAL_DECORATION, BetterBiomePlacedFeatures.MAPLE_PLACED_KEY);
+        biomeBuilder.addFeature(GenerationStep.Decoration.VEGETAL_DECORATION, VegetationPlacements.TREES_PLAINS);
 
         return new Biome.BiomeBuilder()
                 .hasPrecipitation(true)
@@ -64,6 +67,40 @@ public class BetterBiomesBiomes {
                 .build();
     }
 
+    public static Biome thinForest(BootstapContext<Biome> context) {
+        MobSpawnSettings.Builder builder = new MobSpawnSettings.Builder();
+        builder.addSpawn(MobCategory.CREATURE, new MobSpawnSettings.SpawnerData(EntityType.WOLF, 5, 4, 4));
+        BiomeDefaultFeatures.farmAnimals(builder);
+        BiomeDefaultFeatures.commonSpawns(builder);
+
+        BiomeGenerationSettings.Builder biomeBuilder =
+                new BiomeGenerationSettings.Builder(context.lookup(Registries.PLACED_FEATURE),
+                        context.lookup(Registries.CONFIGURED_CARVER));
+
+        globalOverworldGeneration(biomeBuilder);
+        BiomeDefaultFeatures.addDefaultOres(biomeBuilder);
+
+        biomeBuilder.addFeature(GenerationStep.Decoration.VEGETAL_DECORATION, BetterBiomePlacedFeatures.THIN_PLACED_KEY);
+
+        BiomeDefaultFeatures.addForestFlowers(biomeBuilder);
+        BiomeDefaultFeatures.addDefaultExtraVegetation(biomeBuilder);
+
+        return new Biome.BiomeBuilder()
+                .hasPrecipitation(true)
+                .downfall(0.4f)
+                .temperature(0.7f)
+                .generationSettings(biomeBuilder.build())
+                .mobSpawnSettings(builder.build())
+                .specialEffects((new BiomeSpecialEffects.Builder())
+                        .waterColor(0x3f76e4)
+                        .waterFogColor(0x050533)
+                        .skyColor(0x79A6FF)
+                        .grassColorOverride(0x79c05a)
+                        .fogColor(0x93CEF5)
+                        .ambientMoodSound(AmbientMoodSettings.LEGACY_CAVE_SETTINGS)
+                        .build())
+                .build();
+    }
 
 
     public static ResourceKey<Biome> register(String name) {