+++ CHANGELOG.md
@@ -5,3 +5,5 @@ This version brings additions
 - Find this rare structure generating in the desert biome
 - Our tests show there is 1 per ~3000 blocks, so it's a LUCKY find!
 - Re-added REI support
+- Thin Logs and a thin birch forest
+- Maple forest is tweaked to spawn with oak trees